+++ src/get.mjs
@@ -1,52 +1,60 @@
-import {execute, sanitizePath, getOptions, parseValueData} from './util.mjs'
+import {execute, getOptions, parseValueData, debug} from './util.mjs'
 import {extendKeyPath} from './constants.mjs'
 import {Registry} from './Registry.mjs'
+import {VALUE_DEFAULT, VALUE_NOT_SET} from './util.mjs'
 
 
 Registry.prototype.get = function(...args) {
-	return Registry.get(...this._formatArgs(args))
+	return Registry.get(this._formatArgs(args))
 }
 
 
 Registry.prototype.has = async function(...args) {
-	return Registry.has(...this._formatArgs(args))
+	return Registry.has(this._formatArgs(args))
 }
 
 
-Registry.get = function(path, arg2, options) {
-	var type = typeof arg2
-	if (type === 'boolean')
-		return Registry.getKey(path, arg2, options)
-	else if (type === 'object')
-		return Registry.getKey(path, undefined, arg2)
-	else if (type === 'undefined')
-		return Registry.getKey(path)
-	else
-		return Registry.getValue(path, arg2, options)
-}
+Registry.get = async function(...args) {
+	debug('[get.get]', args)
+	var options = getOptions(args)
 
+	// Enable unicode
+	if (options.unicode)
+		var isUnicodeTemporaryEnabled = await Registry.enableUnicode()
 
-Registry.has = function(path, name) {
-	if (name === undefined)
-		return Registry.hasKey(path)
-	else
-		return Registry.hasValue(path, name)
-}
-
-
-Registry.getKey = async function(path, recursive, options) {
-	// Allow both forwardslashes and backslashes
-	path = sanitizePath(path)
-	// [recursive] is optional
-	if (typeof recursive === 'object') {
-		options = recursive
-		recursive = false
+	try {
+		if (options.name !== undefined)
+			var result = await Registry.getValue(options)
+		else
+			var result = await Registry.getKey(options)
+	} finally {
+		// Disable unicode if needed
+		if (options.unicode && isUnicodeTemporaryEnabled)
+			await Registry.disableUnicode()
 	}
-	options = getOptions(options)
-	if (recursive)
-		var result = await execute(['query', path, '/s'])
+
+	return result
+}
+
+Registry.has = function(...args) {
+	debug('[get.has]', args)
+	var options = getOptions(args)
+	if (options.name !== undefined)
+		return Registry.hasValue(options)
 	else
-		var result = await execute(['query', path])
+		return Registry.hasKey(options)
+}
+
+
+Registry.getKey = async function({path, recursive, bitsArg, format, lowercase}) {
+	debug('[get.getKey]', {path, recursive, bitsArg, format, lowercase})
+	if (recursive)
+		var execArgs = ['query', path, '/s']
+	else
+		var execArgs = ['query', path]
+	if (bitsArg)
+	  execArgs.push(bitsArg)
+	var result = await execute(execArgs)
 	// Short circuit further processing if the key at given path was not found and undefined was returned.
 	if (result === undefined) return
 	// Slight preprocessing of the returned result
@@ -57,15 +65,15 @@ Registry.getKey = async function(path, recursive, options) {
 	var trimBy = fullPath.length + 1
 	// The first line might not be equal to path, but 
 	var lastPath = fullPath
-	var root = createKeyStructure(options.format)
+	var root = createKeyStructure(format)
 	var scope = root
 
 	for (var i = 1; i < lines.length; i++) {
 		let line = lines[i]
 		switch (line[0]) {
 			case ' ':
-				let entry = processValueLine(line, options)
-				assignValueEntry(scope, entry, options.format)
+				let entry = processValueLine(line, lowercase)
+				assignValueEntry(scope, entry, format)
 				break
 			case 'H':
 				// Line starts with H, as for HK hive name, process it as subkey path.
@@ -73,9 +81,9 @@ Registry.getKey = async function(path, recursive, options) {
 				var subPath = line.slice(trimBy)
 				if (subPath === '')
 					continue
-				if (options.lowercase)
+				if (lowercase)
 					subPath = subPath.toLowerCase()
-				scope = traverseKeyPath(root, subPath, recursive, options.format)
+				scope = traverseKeyPath(root, subPath, recursive, format)
 				break
 		}
 	}
@@ -86,15 +94,19 @@ Registry.getKey = async function(path, recursive, options) {
 
 // returns single value
 // Warning: 'reg' command is case insensitive and value names in the results are as well.
-Registry.getValue = async function(path, name = Registry.DEFAULT, options) {
-	// Allow both forwardslashes and backslashes
-	path = sanitizePath(path)
-	options = getOptions(options)
+Registry.getValue = async function({path, name, bitsArg, format, lowercase}) {
+	debug('[get.getValue]', {path, name, bitsArg, format, lowercase})
+	// If name is ommited, it will be considered as 'default' registry value
+	if (name === undefined)
+		name = Registry.DEFAULT
 	// Create query for retrieving only single value entry. Either the default one (empty string) or concrete named.
 	if (name === Registry.DEFAULT)
-		var result = await execute(['query', path, '/ve'])
+		var execArgs = ['query', path, '/ve']
 	else
-		var result = await execute(['query', path, '/v', name])
+		var execArgs = ['query', path, '/v', name]
+	if (bitsArg)
+		execArgs.push(bitsArg)
+	var result = await execute(execArgs)
 	// Short circuit further processing if the key at given path was not found and undefined was returned.
 	if (result === undefined) return
 
@@ -103,8 +115,8 @@ Registry.getValue = async function(path, name = Registry.DEFAULT, options) {
 		.split('\r\n')
 		.find(line => line[0] === ' ')
 	// Parse the lino into entry object and return it (or only the data if simple result is requested).
-	var entry = processValueLine(line, options)
-	if (options.format === Registry.FORMAT_SIMPLE)
+	var entry = processValueLine(line, lowercase)
+	if (format === Registry.FORMAT_SIMPLE)
 		return entry.data
 	else
 		return entry
@@ -112,12 +124,13 @@ Registry.getValue = async function(path, name = Registry.DEFAULT, options) {
 
 
 // Returns an array of subkeys (or their full paths).
-Registry.getKeys = async function(path, options) {
-	// Allow both forwardslashes and backslashes
-	path = sanitizePath(path)
-	options = getOptions(options)
+Registry.getKeys = async function({path, bitsArg, format, lowercase}) {
+	debug('[get.getKeys]', {path, bitsArg, format, lowercase})
 	// Create simple query at given path
-	var result = await execute(['query', path])
+	var execArgs = ['query', path]
+	if (bitsArg)
+		execArgs.push(bitsArg)
+	var result = await execute(execArgs)
 	// Short circuit further processing if the key at given path was not found and undefined was returned.
 	if (result === undefined) return
 	// REG can take shorter paths, but returns strictily long paths.
@@ -129,33 +142,34 @@ Registry.getKeys = async function(path, options) {
 		.filter(line => line.startsWith('HK'))
 		.filter(line => line !== fullPath)
 	// simple format trims whole HK.. path and only returns subkeys (the name after last slash).
-	if (options.format === Registry.FORMAT_SIMPLE) {
+	if (format === Registry.FORMAT_SIMPLE) {
 		let trimBy = fullPath.length + 1
 		keyPaths = keyPaths.map(path => path.slice(trimBy))
 	}
 	// Paths and names in registry are case insensity, we can lowercase at this point without loosing any details.
-	if (options.lowercase)
+	if (lowercase)
 		keyPaths = keyPaths.map(path => path.toLowerCase())
 	// return processed paths or subkey names.
 	return keyPaths
 }
 
 
-Registry.getValues = async function(path, options) {
-	// Allow both forwardslashes and backslashes
-	path = sanitizePath(path)
-	options = getOptions(options)
+Registry.getValues = async function({path, bitsArg, format, lowercase}) {
+	debug('[get.getValues]', {path, bitsArg, format, lowercase})
 	// Create simple query at given path
-	var result = await execute(['query', path])
+	var execArgs = ['query', path]
+	if (bitsArg)
+		execArgs.push(bitsArg)
+	var result = await execute(execArgs)
 	// Short circuit further processing if the key at given path was not found and undefined was returned.
 	if (result === undefined) return
 	// Split result by lines and only keep the ones starting with space - only those contain value entry.
 	var entries = result
 		.split('\r\n')
 		.filter(line => line[0] === ' ')
-		.map(line => processValueLine(line, options))
+		.map(line => processValueLine(line, lowercase))
 	// Output in simple or complex format.
-	if (options.format === Registry.FORMAT_SIMPLE) {
+	if (format === Registry.FORMAT_SIMPLE) {
 		var outObject = {}
 		entries.forEach(entry => outObject[entry.name] = entry.data)
 		return outObject
@@ -166,26 +180,32 @@ Registry.getValues = async function(path, options) {
 
 
 // Returns true if a key at the path exists
-Registry.hasKey = function(path) {
-	// Allow both forwardslashes and backslashes
-	path = sanitizePath(path)
+Registry.hasKey = function({path, bitsArg}) {
+	debug('[get.hasKey]', {path, bitsArg})
 	// Create query for retrieving only single value entry. Either the default one (empty string) or concrete named.
+	var execArgs = ['query', path]
+	if (bitsArg)
+		execArgs.push(bitsArg)
 	// 'false' argument disables suppression of not found errors for simpler handling (with catch statement).
-	return execute(['query', path])
+	return execute(execArgs)
 		.then(result => result === undefined ? false : true)
 }
 
 
 // Returns true if a value at the path exists
-Registry.hasValue = async function(path, name = Registry.DEFAULT) {
-	// Allow both forwardslashes and backslashes
-	path = sanitizePath(path)
+Registry.hasValue = async function({path, name, bitsArg}) {
+	debug('[get.hasKey]', {path, name, bitsArg})
+	// If name is ommited, it will be considered as 'default' registry value
+	if (name === undefined)
+		name = Registry.DEFAULT
 	// Create query for retrieving only single value entry. Either the default one (empty string) or concrete named.
-	// 'false' argument disables suppression of not found errors for simpler handling (with catch statement).
 	if (name === Registry.DEFAULT)
-		var result = await execute(['query', path, '/ve'])
+		var execArgs = ['query', path, '/ve']
 	else
-		var result = await execute(['query', path, '/v', name])
+		var execArgs = ['query', path, '/v', name]
+	if (bitsArg)
+		execArgs.push(bitsArg)
+	var result = await execute(execArgs)
 	if (result === undefined) return false
 	// Default value name is represented by a word default in brackets.
 	if (name === Registry.DEFAULT)
@@ -274,7 +294,7 @@ function assignValueEntry(scope, entry, format) {
 }
 
 
-function processValueLine(line, options) {
+function processValueLine(line, lowercase) {
 	// value line starts with 4 spaces and consists of three items that are also spaces by 4 spaces.
 	// WARNING: Do not trim. Lines only start with 4 spaces, they don't end with 4 spaces unsless the value is empty string.
 	var [name, type, data] = line.slice(4).split('    ')
@@ -282,7 +302,7 @@ function processValueLine(line, options) {
 		name = Registry.DEFAULT
 	if (data === VALUE_NOT_SET)
 		data = undefined
-	else if (options.lowercase)
+	else if (lowercase)
 		name = name.toLowerCase()
 	// Convert REG_BINARY value into buffer
 	var [data, type] = parseValueData(data, type)